+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Karafka framework changelog
 
-## Unreleased
+## 2.0.12 (2022-10-06)
 - Commit stored offsets upon rebalance revocation event to reduce number of messages that are re-processed.
 - Support cooperative-sticky rebalance strategy.
 - Replace offset commit after each batch with a per-rebalance commit.